(#11082) Add negative tests for operatingsystemrelease
The current behaviour of the operatingsystemrelease fact is that we have
a general operatingsystemrelease fact that will suit every platform and
returns the kernelrelease fact.

So if our solaris operatingsystemrelease fact does not work
(/etc/release not present, empty, unexpected content) facter will try
to resolve the general operatingsystemrelease fact. That is now
reflected in the tests.