(#11732) Add windows support for ec2 facts
This patch adds support for detecting ec2 on windows. This works by modifying
the linux methodology by using arp -a instead of arp -an and searching for the
mac address with a hyphen delimiter (as apposed to a quote).

I've added tests and a sample fixture which adds output from arp -a from a
windows machine.