(#9574) Simplified fact and adjusted spec tests accordingly
The fact itself was using an 'if' block which duplicated the confine mechanism
we already have for facter. Also the fact was using a mutex for appending to an
array, which was probably unnecessary since the array scope is confined.

I've renamed the spec test to 'filesystems_spec.rb' to reflect the fact name
and moved the fixtures into the appropriate area so my_fixture_read will just
work.